+++ src/main/java/com/Kim/blog/repository/ReplyRepository.java
@@ -3,5 +3,5 @@ package com.Kim.blog.repository;
 import com.Kim.blog.model.Reply;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReplyRepository extends JpaRepository<Reply, Integer> {
+public interface ReplyRepository extends JpaRepository<Reply, Long> {
 }
